HW_React_28.1_v25_'add logic updateUserId'

+++ src/Context.jsx
@@ -47,8 +47,12 @@ export function Provider({children}) {
         setDataUsers(newUsers);
     }
 
+    const updateUserId = (dataId) => {
+        console.log('Hello')
+    }
+
     return (
-        <ContextData.Provider value={{darkMode, toggleTheme, fetchData, users, setUsers, deleteAll, deleteUserId}}>
+        <ContextData.Provider value={{darkMode, toggleTheme, fetchData, users, setUsers, deleteAll, deleteUserId, updateUserId}}>
             {children}
         </ContextData.Provider>
     );
